Changed menu "drop-up" behavior
Existing code makes a menu "drop up" instead of down anytime the menu
bottom would be below the viewport. No consideration is paid to whether
the top of the menu would be visible (or even partially above the
document top, where it becomes inaccessible).

This code change makes it so menus always drop down unless the bottom
would be below the veiewport and the viewport has room to display the
entire menu above the triggering coordinates.